查询考卷api

+++ src/main/java/com/online_exam_sys/pojo/Ex_question.java
@@ -13,5 +13,5 @@ public class Ex_question {
     private Long ep_id;
     private Long qu_id;
     private String eq_answer;
-    private int eq_score;
+    private Float eq_score;
 }
